Move addMinutesToADate helper
